refactor file_dump.py
renamed module and FileDumper class, to reflect their full responsibility: data dumping. Created classes, that handle writing to files with different formats, as well as an interface that  handles interactions with objects of that classes. Inside of DataDumper created a method writer_selector, that returns an object of a specified writer, that corresponds to a file_type - a parameter that DataDumper takes in at init