refactored data_dump.py
the WriterFactory class as it was, wasn't a proper factory, because logic of Writer object choice was still implemented in DataDumper class. DataDumper class was not only implementing the logic to choose the proper format to get a writer, but was also responsible for dumping data into proper writer, which violates SRP. refactored DataDumper takes in a factory object, and when the dump_data method is called with data, output type and output name as parameters, gets appropriate writer object from the factory, and dumping data into it.
Thus, FileDump is responsible only for dumping data into a writer, and WriterFactory is respinsible for providing proper writer wich follows SRP. The FileDump doesn't depend upon specific writer objects, which loosens the cohesion and follows DIP