Create db_operations.py
created db_operations module, that handles everything when it comes to a database. It has a DBConnection class, that can be injected into any high-level function that operates with a db. So far a create_tables was realised, that creates tables rooms and students, corresponding to rooms and students json files